+++ processor/src/test/fixtures/bad_input/com/example/action/JaxWsActionModel.java
@@ -1,0 +1,12 @@
+package com.example.action;
+
+import arez.annotations.ArezComponent;
+
+@ArezComponent( allowEmpty = true )
+public abstract class JaxWsActionModel
+{
+  @javax.xml.ws.Action
+  void setField()
+  {
+  }
+}